requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
chromadb==0.4.18
google-generativeai==0.3.2
python-multipart==0.0.6
python-dotenv==1.0.0
PyPDF2==3.0.1
